Winter 2023 attempt
IPOpt Solver gives some interesting outputs for symmetric shapes like a circle. But, it mostly fails when we move on to rectangles. It is mainly because rectangles have an orientation as well, and we also need to satisfy a lot of complex adjacency relations, which require boolean logic. It is not directly employable in IPOpt.