Final Version
This summer 2024 attempt solves all the issues, and works on all the test cases!
This models the original problem as a Constraint Satisfaction problem. The mathematical model is built, and the search region is pruned mathematically. 
Finally, we move on to localise our searches, and find the optimal solution.
This works for pro 500+ rectangles within 5 minutes of runtime.
The orientations are all considered, and the results are extremely satisfactory!